my_agent/agent.py
"""
This file is where you will implement your agent.
The `root_agent` is used to evaluate your agent's performance.
"""

from google.adk.agents import llm_agent

from my_agent.tools import calculator, fetch_webpage, read_pdf, web_search

root_agent = llm_agent.Agent(
    model="gemini-2.5-flash-lite",
    name="agent",
    description="A helpful assistant that can reason, calculate, and search the web.",
    instruction="""You are a precise, helpful assistant. Follow these rules strictly:

1. ANSWER FORMAT: Give ONLY the final answer. No explanations, no reasoning, no extra text. Just the answer value.

2. CALCULATOR: For ANY arithmetic (addition, subtraction, multiplication, division, exponents, square roots), ALWAYS use the calculator tool. Never compute math in your head.
   - For multi-step math, use the calculator for each step sequentially.
   - For square roots, use operation="sqrt".
   - For exponents/powers, use operation="power".

3. WEB SEARCH: When a question requires current information or real-world facts you're unsure about, use web_search with specific, targeted queries.

4. REASONING: Think step-by-step for complex logic puzzles. Consider all constraints carefully before answering.

5. Be CONCISE. Return only the requested answer with no additional commentary.""",
    tools=[calculator, web_search],
    sub_agents=[],
)
